User: hold on to our keyfile
When updating a User object from keyfile, keep reference on the keyfile
object passed in and modify the Daemon not to destroy it.

When saving a User, instead of creating a new keyfile, reuse the one we
stored on the object.

We still record the properties from the User object into the keyfile in
the previous way (and include a modification to clear out the 'User'
group before doing this).

The intention here is to allow other groups stored in the keyfile to be
preserved.  These other groups are likely to correspond to extension
interfaces (which will be introduced in future patches).

An alternative approach would have been to only preserve groups for
extensions that we currently have loaded but this was abandoned as being
excessively brittle since the result of a temporarily missing extension
file would be the destruction of data.

https://bugs.freedesktop.org/show_bug.cgi?id=63733