user: treat NULL shadow entry as "password handled remotely"
Right now we set the password mode of a user based on the password
hash in shadow.  SSSD (and other remote login facilities
presumably) don't send the password hash over the wire to the client,
but of course those users do have a password set.

This commit changes the code to notice when there's no shadow entry
for a user and treat that case conservatively (assume user has a password).

Related to: https://bugzilla.gnome.org/show_bug.cgi?id=708997

https://bugs.freedesktop.org/show_bug.cgi?id=70005