daemon: call userdel consistently
commit c5905497733bebf9936a7b028a11ca87caf2d71f changed one of
the two calls to userdel in the code to use -f.  This change
was made to ensure users in the logind "closing" state can be
deleted (userdel runs pgrep to make sure a user isn't logged in
unless -f is passed).

Unfortunately, that commit missed the other userdel invocation,
and so now there is inconsistent behavior.

This commit changes makes sure both invocations pass -f.

https://bugs.freedesktop.org/show_bug.cgi?id=65752