daemon: split wtmp code into separate file
Split the code for opening and iterating the wtmp entries into a
separate 'wtmp-helper' file.

For now, this is a direct transplant of the existing code to a new file
with some minimal renaming and a function to get the name of the file to
monitor for changes (which is still setup from daemon.c).