wtmp: stop using private _PATH_WTMPX
Instead of using this private macro, use WTMPX_FILENAME.

Also, deal explicitly with the implications of this define being
missing: throw an #error.

This means that some systems that are currently broken (FreeBSD for
example) will continue to be broken, but will throw an #error about it.

Fixes coming in the following commits.

https://bugs.freedesktop.org/show_bug.cgi?id=75938