daemon: emulate fgetpwent() if we don't have it
We use fgetpwent directly on /etc/passwd in order to ensure we only get
a list of local users (and not ones from the network directory service).
Unfortunately, this function is not commonly found on non-GNU systems.

Provide our own implementation of fgetpwent if the operating system does
not provide it.

https://bugs.freedesktop.org/show_bug.cgi?id=41747