daemon: fix object path derivation for large UIDs
accountsservice exports an object on the system bus
for every user it tracks. The path of the object is
derived from the user's uid.

The format string used for computing the object path
supposes the UID is signed. UIDs are unsigned quantities,
though, so very large UIDs get misrepresented as negative
values in the resulting object path. The negative sign
"-" is invalid in an object path. Under these circumstances,
accountsservice will crash.

This commit corrects the problem, by using a format string
that is appropriate for unsigned values.

https://bugs.freedesktop.org/show_bug.cgi?id=85688