accountsservice: Add SetPasswordHint function
So far the only way of setting the password hint is by setting also the
password.

But there are several cases where setting the password isn't done through
AccountsService, but still we want to set the password hint. For instance GDM
setups the pasword for new used directly from PAM.

Thus, let's add a new function to set the hint without changing the password.

https://bugs.freedesktop.org/show_bug.cgi?id=89588